Moving more code out of evaluator to abstract evaluator
